requirements.txt
ccxt==4.2.15
python-dotenv==1.0.0 